fix: streamBuilder

+++ lib/app/screens/detail/details_screen.dart
@@ -6,7 +6,7 @@ import '../../constants/gasout_constants.dart';
 class DetailsScreen extends StatefulWidget {
   final imgPath;
   final int averageValue;
-  final double maxValue;
+  final int maxValue;
   final int totalHours;
   final String? email;
   final String? roomName;
